Moving cumsum to arithmetics plugin
