+++ phylanx/plugins/arithmetics/cumsum.hpp
