+++ src/plugins/arithmetics/cumsum.cpp
@@ -6,7 +6,7 @@
 #include <phylanx/config.hpp>
 #include <phylanx/ir/node_data.hpp>
 #include <phylanx/execution_tree/primitives/node_data_helpers.hpp>
-#include <phylanx/plugins/matrixops/cumsum.hpp>
+#include <phylanx/plugins/arithmetics/cumsum.hpp>
 
 #include <hpx/include/lcos.hpp>
 #include <hpx/include/naming.hpp>
