+++ tests/unit/plugins/arithmetics/cumsum.cpp
